requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.12
aiosignal==1.3.2
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.3.0
babel==2.16.0
backports.tarfile==1.2.0
billiard==4.2.1
build==1.2.2.post1
CacheControl==0.14.2
cachetools==5.5.2
celery==5.5.2
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
cleo==2.1.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
crashtest==0.4.1
cron-descriptor==1.4.5
dashscope==1.23.4
decorator==5.2.1
diff-match-patch==20241021
distlib==0.3.9
distro==1.9.0
Django==5.1.4
django-celery-beat==2.7.0
django-cors-headers==4.6.0
django-debug-toolbar==4.4.6
django-guardian==2.4.0
django-import-export==4.3.3
django-modeltranslation==0.19.12
django-money==3.5.3
django-simple-history==3.7.0
django-timezone-field==7.1
django-unfold==0.44.0
dnspython==2.7.0
dulwich==0.22.7
exceptiongroup==1.3.0
fastjsonschema==2.21.1
filelock==3.17.0
findpython==0.6.2
frozenlist==1.7.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.0
google-api-python-client==2.171.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.73.0
grpcio-status==1.71.0
gunicorn==22.0.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_metadata==8.6.1
installer==0.7.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jeepney==0.8.0
jiter==0.10.0
keyring==25.6.0
kombu==5.5.3
minio==7.2.15
more-itertools==10.6.0
moviepy==2.2.1
msgpack==1.1.0
multidict==6.4.4
numpy==2.2.6
openai==1.84.0
packaging==24.2
pbs-installer==2025.2.12
pillow==11.1.0
pip==25.0.1
pkginfo==1.12.0
platformdirs==4.3.6
poetry==2.1.1
poetry-core==2.1.1
proglog==0.1.12
prompt_toolkit==3.0.48
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
psycopg==3.2.4
psycopg2-binary==2.9.10
py-moneyed==3.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycryptodome==3.22.0
pydantic==2.11.5
pydantic_core==2.33.2
Pygments==2.19.1
PyJWT==2.10.1
pymongo==4.13.2
pyparsing==3.2.3
pyproject_hooks==1.2.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
RapidFuzz==3.12.1
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9.1
SecretStorage==3.3.3
sentry-sdk==2.19.2
setuptools==75.8.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tablib==3.7.0
tomli==2.2.1
tomlkit==0.13.2
tqdm==4.67.1
trove-classifiers==2025.1.15.22
typing_extensions==4.12.2
typing-inspection==0.4.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.3.0
vine==5.1.0
virtualenv==20.29.2
wcwidth==0.2.13
websocket-client==1.8.0
whitenoise==6.8.2
yarl==1.20.1
zipp==3.21.0
zstandard==0.23.0
djangorestframework
upstash-vector
pyvips
tifffile  

